Older_code/requirements.txt
# CrewAI and AI frameworks
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0
shap>=0.42.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework
streamlit>=1.25.0
flask>=2.3.0
dash>=2.11.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Data Handling
pydantic>=2.0.0
sqlalchemy>=2.0.0
pymongo>=4.4.0
redis>=4.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
jsonschema>=4.18.0
click>=8.1.0
tqdm>=4.65.0
joblib>=1.3.0

# Healthcare specific
faker>=19.0.0
scipy>=1.11.0
statsmodels>=0.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0
